+++ tests/prototype.rs
@@ -1,0 +1,13 @@
+//! Use this file for iterating on the derive code. You can view the
+//! expanded code for any given configuration by updating this file and
+//! running:
+//!
+//! ```sh
+//! cargo expand --test=prototype
+//! ```
+
+
+#[test_log::test]
+fn it_works() {
+  assert_eq!(2 + 2, 4);
+}